changed the build tool from maven to gradle for better readability
