+++ .idea/compiler.xml
@@ -2,30 +2,20 @@
 <project version="4">
   <component name="CompilerConfiguration">
     <annotationProcessing>
-      <profile name="Annotation profile for gateway" enabled="true">
-        <sourceOutputDir name="target/generated-sources/annotations" />
-        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
+      <profile name="Gradle Imported" enabled="true">
         <outputRelativeToContentRoot value="true" />
         <processorPath useClasspath="false">
-          <entry name="$MAVEN_REPOSITORY$/org/projectlombok/lombok/1.18.36/lombok-1.18.36.jar" />
-          <entry name="$MAVEN_REPOSITORY$/org/projectlombok/lombok/1.18.36/lombok-1.18.36.jar" />
+          <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.36/5a30490a6e14977d97d9c73c924c1f1b5311ea95/lombok-1.18.36.jar" />
         </processorPath>
-        <module name="gateway" />
-      </profile>
-      <profile name="Annotation profile for ssoservice" enabled="true">
-        <sourceOutputDir name="target/generated-sources/annotations" />
-        <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
-        <outputRelativeToContentRoot value="true" />
-        <processorPath useClasspath="false">
-          <entry name="$MAVEN_REPOSITORY$/org/projectlombok/lombok/1.18.36/lombok-1.18.36.jar" />
-        </processorPath>
-        <module name="ssoservice" />
+        <module name="SsoService.main" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
       <module name="authservice" target="21" />
       <module name="gateway2" target="21" />
+      <module name="gymuserservice" target="21" />
       <module name="loginService" target="21" />
+      <module name="ssoservice2" target="21" />
     </bytecodeTargetLevel>
   </component>
   <component name="JavacSettings">
@@ -33,8 +23,9 @@
       <module name="authservice" options="-parameters" />
       <module name="gateway" options="-parameters" />
       <module name="gateway2" options="-parameters" />
+      <module name="gymuserservice" options="-parameters" />
       <module name="loginService" options="-parameters" />
-      <module name="ssoservice" options="-parameters" />
+      <module name="ssoservice2" options="-parameters" />
     </option>
   </component>
 </project>
