+++ .idea/encodings.xml
@@ -2,8 +2,9 @@
 <project version="4">
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/backend/authService/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/backend/gateway/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/backend/gateway2/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/backend/ssoservice/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/backend/gatewaytest/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/backend/gymuserservice/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/backend/ssoservice2/src/main/java" charset="UTF-8" />
   </component>
 </project>
