+++ .idea/sonarlint.xml
@@ -3,8 +3,11 @@
   <component name="SonarLintProjectSettings">
     <option name="moduleMapping">
       <map>
+        <entry key="SsoService" value="SsoService" />
         <entry key="gateway2" value="gateway" />
+        <entry key="gymuserservice" value="gymuser" />
         <entry key="loginService" value="authService" />
+        <entry key="ssoservice2" value="ssoservice" />
       </map>
     </option>
   </component>
