+++ backend/ssoService/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'SsoService'
