+++ backend/ssoService/src/main/java/com/nextlift/SsoService/SsoServiceApplication.java
@@ -1,0 +1,13 @@
+package com.nextlift.SsoService;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SsoServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SsoServiceApplication.class, args);
+	}
+
+}
