+++ backend/ssoService/src/main/java/com/nextlift/SsoService/controller/GymUserController.java
@@ -1,0 +1,4 @@
+package com.nextlift.SsoService.controller;
+
+public class GymUserController {
+}
