+++ backend/ssoService/src/main/java/com/nextlift/SsoService/model/GymUser.java
@@ -1,0 +1,16 @@
+package com.nextlift.SsoService.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class GymUser {
+
+    private int age;
+    private String name;
+}
