+++ backend/ssoService/src/main/java/com/nextlift/SsoService/repository/GymUserRepository.java
@@ -1,0 +1,4 @@
+package com.nextlift.SsoService.repository;
+
+public interface GymUserRepository {
+}
