+++ backend/ssoService/src/main/java/com/nextlift/SsoService/service/GymUserService.java
@@ -1,0 +1,4 @@
+package com.nextlift.SsoService.service;
+
+public class GymUserService {
+}
