+++ backend/ssoService/src/main/resources/application.yaml
