+++ backend/ssoService/src/test/java/com/nextlift/SsoService/SsoServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.nextlift.SsoService;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SsoServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
