--- backend/ssoservice/.gitattributes
@@ -1,2 +1,0 @@
-/mvnw text eol=lf
-*.cmd text eol=crlf
