--- backend/ssoservice/src/main/java/com/nextlift/ssoservice/SsoServiceApplication.java
@@ -1,13 +1,0 @@
-package com.nextlift.ssoservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SsoServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SsoServiceApplication.class, args);
-	}
-
-}
