--- backend/ssoservice/src/main/java/com/nextlift/ssoservice/controller/GymUserController.java
@@ -1,4 +1,0 @@
-package com.nextlift.ssoservice.controller;
-
-public class GymUserController {
-}
