--- backend/ssoservice/src/main/java/com/nextlift/ssoservice/model/GymUser.java
@@ -1,4 +1,0 @@
-package com.nextlift.ssoservice.model;
-
-public class GymUser {
-}
