--- backend/ssoservice/src/main/java/com/nextlift/ssoservice/repository/GymUserRepository.java
@@ -1,4 +1,0 @@
-package com.nextlift.ssoservice.repository;
-
-public interface GymUserRepository {
-}
