--- backend/ssoservice/src/main/java/com/nextlift/ssoservice/service/GymUserService.java
@@ -1,4 +1,0 @@
-package com.nextlift.ssoservice.service;
-
-public class GymUserService {
-}
