--- backend/ssoservice/src/test/java/com/nextlift/ssoservice/SsoserviceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.nextlift.ssoservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SsoserviceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
